updated to include proper selection of encryption keys through the wizard and on image save. (still with apg)

+++ jni/tiff_ifd.h
@@ -88,7 +88,7 @@ public:
       return NULL;
     return(tags_[tagindex]);
   }
-  int GetNTags() const { return tags_.size(); }
+  size_t GetNTags() const { return tags_.size(); }
   // Print all tags to stdout.
   void Print() const;
   Jpeg *GetJpeg() { return jpeg_;}
